adds encryption to the user password
