+++ src/auth/local/index.ts
@@ -1,0 +1,9 @@
+import { Router } from "express";
+
+import { signUpController } from "./local.controller";
+
+const router = Router();
+
+router.post("/signup", signUpController);
+
+export default router;
